lint: first working version
Added the skeleton of the lint package with the first version of the
schema. The schema originates from an old version we had in another
repo, so it is missing some fields. That will be improved in the future,
when we disallow additional (unknown) fields.

The command-line interface is simple - just run "tyk lint", and the
warnings and any error (if any) will be shown.

Updates #336.