session: decouple lifetime func from APISpec
This will help place the file in a separate package. While at it, reuse
Hash in SetFirstSeenHash. The only difference with that change is that
we may set firstSeenHash to empty string if there is an error, but this
is fine since we only ever call SetFirstSeenHash on new sessions, which
already have firstSeenHash as empty string.