logging output level available as config option
log_level field added to Tyk configuration.

If TYK_LOGLEVEL enviroment variable is unset and no --debug flag is used when executing Tyk then the configuration file will be searched for an alternative to the default info level logging output. The log_level value can be error, warn or debug; Info is redundant as that is the default logger output set in log/log.go. Fixes #1182