analytics: don't panic on nil response copy
This was a regression introduced by myself in d847b5e3. I was wrong in
saying that these were never nil. In one scenario, responseCopy can be
nil - see the added comments.

Perhaps having a non-nil response copy here would be a better approach,
but do the safer revert for now and leave it as a TODO for a later time.

Fixes #1327.