linter guaranteed to exit (#1884)
Addresses one of issues in #1880
I couldnt consistently reproduce the behaviour but sometime when tyk lint is run tyk process will start up afterwards. This is not the advertised behaviour of the linter currently - so this change guarantees exit.

We can change behaviour by attaching flags to the command at a future time.