Fix whitelist/blacklist regex too broad (#1956)
Fixes https://github.com/TykTechnologies/tyk/issues/1944

```
apiLangIDsRegex := regexp.MustCompile(`{([^}]*)}`)
asRegexStr := apiLangIDsRegex.ReplaceAllString(stringSpec, `([^/]*)`)
```

Basically removing logic of detection end of string. And at the same time modifying tests to do that:
```
v.Paths.WhiteList = []string{"/foo/{fooId}$", "/foo/{fooId}/bar/{barId}$"}
```

Basic idea here that here at Tyk we have backward compatibility policy, and this regexp was here for a few years, so there is chance that some of our customers relied on its "confusing" behavior. Thankfuly it was possible to fix it using "$" at the end of string, to explicitly set end of the string. In case of #1944  bug this trick was not working for urls containing {...}. Goal of this PR in this case can be making {...} groups support addign "$" to the end of URL.