basic auth cache - thundering heard (#2323)
resolves: https://github.com/TykTechnologies/tyk/issues/2324

Implemented singleflight, which suppresses/coalesces duplicate BCrypt function
calls for the same cache key in event of cache miss.

The purpose of Caching, is to avoid performing expensive computation if the computation has already been calculated.

But if the cache is empty (cold start, reload, expiry), and under high load, the first `x` requests will get a cache miss, and all simultaneously perform BCrypt + Caching of exactly the same password - then race to fill the cache.

So we have various problems under load… 

1. BCrypt is being called`x-1` too many times, 
2. Necessity to check & set in cache `x-1` too many times.

---

After receiving a cache-miss, we performs BCrypt + Cache via singleflight. Singleflight coalesces similar requests into a single call to BCrypt+Cache function, and returns the result to all of the remaining requests.

As such, when the cache is cleared (hot reload, cold-start, cache expire), we have mitigated the thundering herd problem, because only one of the herd will actually perform the expensive operation, while the others simply wait for the result.