Fix plugin building (#2810)
After the suggestion of keeping vendor'd code in `$GOPATH/src/vendor`, basically this PR boils down to setting GOPATH in the Dockerfile. This was the omission that took me a long time to figure out.

This has been tested end-to-end using the unstable pipeline. The commit that was tested is `dac95f7e04de2ebf63cfc15f555650153233c6dc`. Additionally, to make this reproducible, modules are not fetched and it is assumed that everything required is vendor'd. This should help reproducibility a good deal.