add TestChecker_triggerSampleLimit (#2862)
This adds a test for uptime tests failure_trigger_sample_size option

closes #2698

The debug logs are correct. Its just .a misunderstanding on how this
option work. If failure_trigger_sample_size=3 if the check fails 3
consecutive times one event is fired of host down.

there is a difference  between logs and event firing.

This doesn't mean the check loop is stopped, the check will continue
but events will only fire based on the limit.