handle properly the value scope_allowance on keys  (#2956)
This PR was created in order to give a fix to https://github.com/TykTechnologies/tyk-analytics/issues/1698 and set the proper allowance_scope when is applied a policy with limits set for one api (but 2 apis are in accessRights array) and set policy_Id as `AllowanceScope = policyID` as well as `Limit.SetBy = resourceForScope`. Was added a test as well.