Fix custom keys in MDCB (#3129)
<!-- Provide a general summary of your changes in the Title above -->

## Description
When users attempts to create a custom key, then we should allow them to use this key in any of the slaves nodes. So, at the moment of retriving a key we should handle as well the situation of customs keys, therefore we should check the existence of that key by searching for `generateToken(orgID, keyName) `

In other hand, as we might use a key by alias or by token value then the slave nodes were storing the api based with the value that we used, this is not correct at all as we need to keep consistency in the values stored in redis (to avoid issues on update and delete operations). Basically if we used the alias to generate the hash value (if enabled) and then store that in redis, later if we used the token instead the alias we again hash the value (if enabled) and store that in redis, that made us having duplicates of the same key, in order to fix that then we now consider the option that the key can be modified inside `CheckSessionAndIdentityForValidKey` therefore now that function will receive a `*string` instead `string` (refer to https://github.com/TykTechnologies/tyk/pull/3129/files#diff-76df41fc5dc0183d026b785861779d24R688 )

## Related Issue
https://github.com/TykTechnologies/tyk/issues/3103

## Motivation and Context
Give solution to https://github.com/TykTechnologies/tyk/issues/3103 so custom keys works in MDCB environment. 

## How This Has Been Tested
1- Setup MDCB environment
2- Create API and Policy. Set API with authentication token method
2- Create custom key in Master GW...something like:
```
curl --location --request POST 'http://tyk-gateway:8080/tyk/keys/mycustomkey' \
--header 'x-tyk-authorization: 352d20ee67be67f6340b4c0605b044b7' \
--header 'Content-Type: application/json' \
--data-raw '{
   "allowance":1000,
   "rate":1000,
   "per":60,
   "quota_max":-1,
   "quota_renews":1552996743,
   "quota_remaining":-1,
   "quota_renewal_rate":-1,
   "org_id": "5e2091c4d4aefce60c04fb92",
   "apply_policies":[
      "5ed7cb95d4aefc2e388ecb2a"
   ]
}'
```
3- Consume API in slave node using `mycustomkey` as Authorization
4- You should be allowed to consume API. Eg:
```
curl --location --request GET 'http://tyk-gateway:8182/api-custom-key/' \
--header 'Authorization: mycustomkey'
```
5- As an extra step, we can check the key details doing something like:
```
curl --location --request GET 'tyk-gateway:8080/tyk/keys/mycustomkey?username=true&org_id=5e2091c4d4aefce60c04fb92' \
--header 'X-Tyk-Authorization: 352d20ee67be67f6340b4c0605b044b7'
```

## Screenshots (if appropriate)

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
<!-- Go over all the following points, and put an `x` in all the boxes that apply -->
<!-- If you're unsure about any of these, don't hesitate to ask; we're here to help! -->
- [x] Make sure you are requesting to **pull a topic/feature/bugfix branch** (right side). If pulling from your own
      fork, don't request your `master`!
- [x] Make sure you are making a pull request against the **`master` branch** (left side). Also, you should start
      *your branch* off *our latest `master`*.
- [ ] My change requires a change to the documentation.
  - [ ] If you've changed APIs, describe what needs to be updated in the documentation.
  - [ ] If new config option added, ensure that it can be set via ENV variable
- [ ] I have updated the documentation accordingly.
- [ ] Modules and vendor dependencies have been updated; run `go mod tidy && go mod vendor`
- [ ] When updating library version must provide reason/explanation for this update.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.
- [x] Check your code additions will not fail linting checks:
  - [x] `go fmt -s`
  - [x] `go vet`