Sync wf-gen from tyk-ci using pr.zsh (#3423)
# Description
These related PRs,

Fix the problem with quotes that was seen in https://github.com/TykTechnologies/tyk-analytics/runs/1510100556?check_suite_focus=true

Add a new workflow triggered on branch delete that will help with cleanup of developer environments.

Add support for branches prefixed with `feature/` to get the same benefits as `integration/`.

Add tyk-sink (mdcb) to the integration build.

# Testing
Syntax checks have passed. Most of the real functionality is external to the workflow.

The status of the integration workflows can be inspected in this PR itself in the _Checks_ tab. The branch is prefixed with `feature/` to test that it triggers the required workflows when it is deleted.

# Merging
These changes have to be sync'd to every release branch. tykbot will be used after this is merged to master.