Add allowance scope in case no pol in key and per api (#3586)
The problem is that when you add two APIs to a key without policy, they are using same redis key for quota. In policy case, it uses allowance scope and for each API, it stores in different redis key. 

This is how we generate redis key:
https://github.com/TykTechnologies/tyk/blob/047733bd92e150c110be0d3afa8c2a4106ea3699/gateway/session_manager.go#L257

Briefly, when two APIs key without policy and per api enabled `quotaScope` is `""`, that is the issue.